updates and refactor
