+++ .github/FUNDING.yml
@@ -1,0 +1,1 @@
+github: julianstark999
