+++ .github/SECURITY.md
@@ -1,0 +1,3 @@
+# Security Policy
+
+If you discover any security related issues, please email julianstark1@googlemail.com instead of using the issue tracker.
