+++ composer.json
@@ -24,12 +24,14 @@
         "php": "^7.3|^8.0",
         "illuminate/console": "^8.0",
         "illuminate/database": "^8.0",
-        "illuminate/support": "^8.0"
+        "illuminate/support": "^8.0",
+        "spatie/laravel-package-tools": "^1.1"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.17",
         "phpunit/phpunit": "^9.0",
-        "orchestra/testbench": "^6.0"
+        "orchestra/testbench": "^6.0",
+        "vimeo/psalm": "^4.6"
     },
     "autoload": {
         "psr-4": {
@@ -39,13 +41,6 @@
             "src/helpers.php"
         ]
     },
-    "extra": {
-        "laravel": {
-            "providers": [
-                "JulianStark999\\LaravelModelIid\\LaravelModelIidServiceProvider"
-            ]
-        }
-    },
     "autoload-dev": {
         "psr-4": {
             "JulianStark999\\LaravelModelIid\\Tests\\": "tests"
@@ -54,9 +49,19 @@
     "scripts": {
         "test": "vendor/bin/phpunit",
         "test-coverage": "vendor/bin/phpunit --coverage-html coverage",
-        "lint": "php-cs-fixer fix -v"
+        "lint": "php-cs-fixer fix -v",
+        "psalm": "psalm"
     },
     "config": {
         "sort-packages": true
-    }
+    },
+    "extra": {
+        "laravel": {
+            "providers": [
+                "JulianStark999\\LaravelModelIid\\LaravelModelIidServiceProvider"
+            ]
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
